Implement efficient updating using Butler (and more!) (#32)
Butler is an implementation of the Wharf specification by itch.io. This can effectively compute a diff for an entire directory, by hashing it in 128KB chunks and comparing these chunks to a newer version (which also allows it to easily track moved or renamed files, or files contained within VPKs).

With this PR, the previous updater has been almost completely rewritten to use Butler instead. The remote JSON has been reformatted to accommodate these changes, and as such, the SOURCE_URL now points towards https://wiki.tf2classic.com/downloader/butler (may eventually be symlinked back to /downloader/)

The decision tree has been replaced with a "main menu" where the user has to enter a number. A new feature is implemented through Butler where the user's local files can be verified and repaired. This is done automatically before every upgrade, to make sure patches apply cleanly.

Upgrades are atomic, they're decompressed to a staging directory before finally being moved on top of the current game. This can require a lot of temporary storage, hence the user can now be prompted to select a temporary folder instead of always using the OS default. This uses the Tk folder picker, which is ugly on Linux but more than sufficient on Windows.

Aria2 is now included in Linux builds, making it work on the Steam Deck, or on systems that don't compile Aria2 with Metalink support. With the inclusion of Butler and Tk, the binary size has ballooned to 40MB or so, even with new exclusions to try to remove non-essential modules. Advice on reducing this would be appreciated.

Beyond that, there's just been a lot of internal refactoring and bugfixing. Testing has not been comprehensive and there may still be edge-cases where I've introduced new problems. Particularly, I haven't even looked at script mode yet. Still, I'm hoping to get this good enough to ship soon, so that we can get some public testing before it has to hold up to the weight of 2.1.


---------

Co-authored-by: ChloeC <chloecormier@riseup.net>